File refactoring and major rewrite of grid utilities
I split the files into the mv-grid and mv-gsll directories.

mv-grid contains utilities that will perform only on grids (foreign
and cl arrays) and will not depend or enhance gsll.

The mv-grid utilities include
 - grid generation utilities,
 - grid mapping syntactic suggar with a simpler interface
 - mathematical functions on grids

Read the README file in that directory for status and future directions.

mv-gsll contains gsll related utilities.  They may depend on mv-grid.

The mv-gsll utilities are considered
broken at this time because of large changes in gsll.  They need to be
rechecked and possibly rewritten.